requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
anyio==4.6.2.post1
    # via httpx
attrs==24.2.0
    # via
    #   -r requirements.in
    #   cattrs
    #   outcome
    #   trio
cattrs==24.1.2
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via typer
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.27.2
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   trio
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
outcome==1.3.0.post0
    # via trio
pygments==2.18.0
    # via rich
rich==13.9.4
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   trio
sortedcontainers==2.4.0
    # via trio
trio==0.27.0
    # via -r requirements.in
typer==0.13.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via typer
